feat: place baseQuery in separate module
